Add mirror implementation of Segmentation in C++/CLI
Provides a base for migration of implementation into native code.
For now, two constructors had to be made public due to need
for visibility in C++ code. They should be made internal after whole
process or even removed.
Unnecessary code was removed from Segmentation as it was not linked
to further processing and caused no error.