Add tests for new functionalities
Provides some testing for MATLAB interfaces extracted outside.

Interfaces for removal of redundant components are turned off now, since
no one knows, how the filtration threshold was set in MATLAB and therefore
it is impossible to specify it in the code. When we gain this knowledge,
it must be specified in MATLAB code and libraries must be recompiled.
Until that time, NotImplementedException will be thrown.