Introduce ReachedUnreachableCodeException
It is enforced to prevent compiler from issuing a warning, that a path
exists without a return. This way we may mark such a path.
Thrown when we enter region we know it is impossible to enter.
This means that assumptions regarding the code design changed
and code should be revised in much more detailed manner.